refactor: move jsonschema stuff again
